db関係ファイルの移動 (#139)
