+++ api/migrate_db.py
@@ -1,6 +1,7 @@
 from sqlalchemy import create_engine
 
-from v2.mysql.model import Base
+from db import Base
+from v2.models import Couple, Message, Text, User  # noqa: F401
 from v2.utils.config import get_db_url
 
 DB_URL = get_db_url()
