+++ api/v2/models/__init__.py
@@ -1,0 +1,5 @@
+# flake8: noqa
+from .couple import Couple
+from .message import Message
+from .text import Text
+from .user import User
