+++ api/v2/models/couple.py
@@ -1,0 +1,20 @@
+from sqlalchemy import TIMESTAMP, Column, Integer, String
+from sqlalchemy.orm import relationship
+from sqlalchemy.sql import func
+
+from db import Base
+
+
+class Couple(Base):
+    __tablename__ = "couples"
+
+    id = Column(Integer, primary_key=True, autoincrement=True)
+    name = Column(String(20))
+    created_at = Column(TIMESTAMP, server_default=func.current_timestamp())
+    updated_at = Column(
+        TIMESTAMP,
+        server_default=func.current_timestamp(),
+        onupdate=func.current_timestamp(),
+    )
+
+    users = relationship("User", back_populates="couple")
