+++ api/v2/models/message.py
@@ -1,0 +1,14 @@
+from sqlalchemy import Column, ForeignKey, Integer, String
+from sqlalchemy.orm import relationship
+
+from db import Base
+
+
+class Message(Base):
+    __tablename__ = "messages"
+
+    id = Column(Integer, primary_key=True, autoincrement=True)
+    user_id = Column(Integer, ForeignKey("users.id"))
+    send_message = Column(String(1000))
+
+    user = relationship("User", back_populates="messages")
