+++ api/v2/models/text.py
@@ -1,0 +1,17 @@
+from sqlalchemy import TIMESTAMP, Column, ForeignKey, Integer, String
+from sqlalchemy.orm import relationship
+from sqlalchemy.sql import func
+
+from db import Base
+
+
+class Text(Base):
+    __tablename__ = "texts"
+
+    id = Column(Integer, primary_key=True, autoincrement=True)
+    user_id = Column(Integer, ForeignKey("users.id"))
+    prompt = Column(String(1000))
+    generated_text = Column(String(1000))
+    created_at = Column(TIMESTAMP, server_default=func.current_timestamp())
+
+    user = relationship("User", back_populates="texts")
