+++ api/v2/models/user.py
@@ -1,0 +1,19 @@
+from sqlalchemy import TIMESTAMP, Column, ForeignKey, Integer, String
+from sqlalchemy.orm import relationship
+from sqlalchemy.sql import func
+
+from db import Base
+
+
+class User(Base):
+    __tablename__ = "users"
+
+    id = Column(Integer, primary_key=True, autoincrement=True)
+    couple_id = Column(Integer, ForeignKey("couples.id"))
+    name = Column(String(20))
+    created_at = Column(TIMESTAMP, server_default=func.now())
+    updated_at = Column(TIMESTAMP, server_default=func.now(), onupdate=func.now())
+
+    couple = relationship("Couple", back_populates="couples")
+    texts = relationship("Text", back_populates="user")
+    messages = relationship("Message", back_populates="user")
